+++ server/index.ts
@@ -4,6 +4,7 @@ import { TextDocuments } from 'vscode-languageserver/node';
 import { registerHandlers } from './src/handlers/handlers';
 import { initializeLogger, log, logErrorWithStack } from './src/utils/logger';
 import { DocumentationService } from './src/services/documentation-service';
+import { CompletionService } from './src/services/completion-service';
 
 const connection = createConnection(ProposedFeatures.all);
 const documents = new TextDocuments(TextDocument);
@@ -20,6 +21,7 @@ process.on('unhandledRejection', (reason) => {
 
 // Initialize services
 const documentationService = new DocumentationService();
+const completionService = new CompletionService(documentationService);
 
 documents.listen(connection);
 
@@ -27,6 +29,7 @@ registerHandlers({
     connection,
     documents,
     documentationService,
+    completionService,
 });
 
 connection.onInitialized(() => {